+++ src/main/java/com/deare/backend/api/auth/dto/response/TermResponseDTO.java
@@ -2,7 +2,6 @@ package com.deare.backend.api.auth.dto.response;
 
 import com.deare.backend.domain.term.entity.Term;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 /**
@@ -11,28 +10,6 @@ import java.util.List;
 public record TermResponseDTO(
         List<TermItemDTO> terms
 ) {
-    public record TermItemDTO(
-            Long termId,
-            String title,
-            String type,
-            String content,
-            boolean isRequired,
-            LocalDateTime effectiveAt,
-            String version
-    ) {
-        public static TermItemDTO from(Term term) {
-            return new TermItemDTO(
-                    term.getId(),
-                    term.getTitle(),
-                    term.getType().name(),
-                    term.getContent(),
-                    term.isRequired(),
-                    term.getEffectiveAt(),
-                    term.getVersion()
-            );
-        }
-    }
-
     public static TermResponseDTO from(List<Term> terms) {
         List<TermItemDTO> termItems = terms.stream()
                 .map(TermItemDTO::from)
